+++ app/features/calendar/actions/calendar.new.server.ts
@@ -1,21 +1,16 @@
 import type { ActionFunction } from "@remix-run/node";
 import { redirect } from "@remix-run/node";
-import { z } from "zod";
-import {
-	type CalendarEventTag,
-	type PersistedCalendarEventTag,
-	TOURNAMENT_STAGE_TYPES,
-} from "~/db/tables";
+import type { CalendarEventTag } from "~/db/tables";
 import { requireUser } from "~/features/auth/core/user.server";
 import * as CalendarRepository from "~/features/calendar/CalendarRepository.server";
+import { newCalendarEventActionSchema } from "~/features/calendar/calendar-schemas";
 import * as ShowcaseTournaments from "~/features/front-page/core/ShowcaseTournaments.server";
 import { MapPool } from "~/features/map-list-generator/core/map-pool";
-import * as Progression from "~/features/tournament-bracket/core/Progression";
+import { notify } from "~/features/notifications/core/notify.server";
 import {
 	clearTournamentDataCache,
 	tournamentFromDB,
 } from "~/features/tournament-bracket/core/Tournament.server";
-import { TOURNAMENT } from "~/features/tournament/tournament-constants";
 import { rankedModesShort } from "~/modules/in-game-lists/modes";
 import { requireRole } from "~/modules/permissions/guards.server";
 import {
@@ -29,24 +24,9 @@ import {
 	uploadImageIfSubmitted,
 } from "~/utils/remix.server";
 import { calendarEventPage } from "~/utils/urls";
-import {
-	actualNumber,
-	checkboxValueToBoolean,
-	date,
-	falsyToNull,
-	id,
-	processMany,
-	removeDuplicates,
-	safeJSONParse,
-	toArray,
-} from "~/utils/zod";
-import { CALENDAR_EVENT, REG_CLOSES_AT_OPTIONS } from "../calendar-constants";
+import { CALENDAR_EVENT } from "../calendar-constants";
 import { canEditCalendarEvent } from "../calendar-utils";
-import {
-	calendarEventMaxDate,
-	calendarEventMinDate,
-	regClosesAtDate,
-} from "../calendar-utils";
+import { regClosesAtDate } from "../calendar-utils";
 
 export const action: ActionFunction = async ({ request }) => {
 	const user = await requireUser(request);
@@ -96,6 +76,7 @@ export const action: ActionFunction = async ({ request }) => {
 		bracketProgression: data.bracketProgression ?? null,
 		minMembersPerTeam: data.minMembersPerTeam ?? undefined,
 		isRanked: data.isRanked ?? undefined,
+		isTest: data.isTest ?? undefined,
 		isInvitational: data.isInvitational ?? false,
 		deadlines: data.strictDeadline ? ("STRICT" as const) : ("DEFAULT" as const),
 		enableNoScreenToggle: data.enableNoScreenToggle ?? undefined,
@@ -177,145 +158,21 @@ export const action: ActionFunction = async ({ request }) => {
 		clearTournamentDataCache(createdTournamentId);
 		ShowcaseTournaments.clearParticipationInfoMap();
 		ShowcaseTournaments.clearCachedTournaments();
+
+		if (data.isTest) {
+			notify({
+				notification: {
+					type: "TO_TEST_CREATED",
+					meta: {
+						tournamentName: data.name,
+						tournamentId: createdTournamentId,
+					},
+				},
+				defaultSeenUserIds: [user.id],
+				userIds: [user.id],
+			});
+		}
 	}
 
 	throw redirect(calendarEventPage(createdEventId));
 };
-
-export const bracketProgressionSchema = z.preprocess(
-	safeJSONParse,
-	z
-		.array(
-			z.object({
-				type: z.enum(TOURNAMENT_STAGE_TYPES),
-				name: z.string().min(1).max(TOURNAMENT.BRACKET_NAME_MAX_LENGTH),
-				settings: z.object({
-					thirdPlaceMatch: z.boolean().optional(),
-					teamsPerGroup: z.number().int().optional(),
-					groupCount: z.number().int().optional(),
-					roundCount: z.number().int().optional(),
-				}),
-				requiresCheckIn: z.boolean(),
-				startTime: z.number().optional(),
-				sources: z
-					.array(
-						z.object({
-							bracketIdx: z.number(),
-							placements: z.array(z.number()),
-						}),
-					)
-					.optional(),
-			}),
-		)
-		.refine(
-			(progression) =>
-				Progression.bracketsToValidationError(progression) === null,
-			"Invalid bracket progression",
-		),
-);
-
-export const calendarEventTagSchema = z
-	.string()
-	.refine((val) =>
-		CALENDAR_EVENT.PERSISTED_TAGS.includes(val as PersistedCalendarEventTag),
-	);
-
-export const newCalendarEventActionSchema = z
-	.object({
-		eventToEditId: z.preprocess(actualNumber, id.nullish()),
-		tournamentToCopyId: z.preprocess(actualNumber, id.nullish()),
-		organizationId: z.preprocess(actualNumber, id.nullish()),
-		name: z
-			.string()
-			.min(CALENDAR_EVENT.NAME_MIN_LENGTH)
-			.max(CALENDAR_EVENT.NAME_MAX_LENGTH),
-		description: z.preprocess(
-			falsyToNull,
-			z.string().max(CALENDAR_EVENT.DESCRIPTION_MAX_LENGTH).nullable(),
-		),
-		rules: z.preprocess(
-			falsyToNull,
-			z.string().max(CALENDAR_EVENT.RULES_MAX_LENGTH).nullable(),
-		),
-		date: z.preprocess(
-			toArray,
-			z
-				.array(
-					z.preprocess(
-						date,
-						z.date().min(calendarEventMinDate()).max(calendarEventMaxDate()),
-					),
-				)
-				.min(1)
-				.max(CALENDAR_EVENT.MAX_AMOUNT_OF_DATES),
-		),
-		bracketUrl: z
-			.string()
-			.url()
-			.max(CALENDAR_EVENT.BRACKET_URL_MAX_LENGTH)
-			.default("https://sendou.ink"),
-		discordInviteCode: z.preprocess(
-			falsyToNull,
-			z.string().max(CALENDAR_EVENT.DISCORD_INVITE_CODE_MAX_LENGTH).nullable(),
-		),
-		tags: z.preprocess(
-			processMany(safeJSONParse, removeDuplicates),
-			z.array(calendarEventTagSchema).nullable(),
-		),
-		badges: z.preprocess(
-			processMany(safeJSONParse, removeDuplicates),
-			z.array(id).nullable(),
-		),
-		avatarImgId: id.nullish(),
-		pool: z.string().optional(),
-		toToolsEnabled: z.preprocess(checkboxValueToBoolean, z.boolean()),
-		toToolsMode: z.enum(["ALL", "TO", "SZ", "TC", "RM", "CB"]).optional(),
-		isRanked: z.preprocess(checkboxValueToBoolean, z.boolean().nullish()),
-		regClosesAt: z.enum(REG_CLOSES_AT_OPTIONS).nullish(),
-		enableNoScreenToggle: z.preprocess(
-			checkboxValueToBoolean,
-			z.boolean().nullish(),
-		),
-		enableSubs: z.preprocess(checkboxValueToBoolean, z.boolean().nullish()),
-		autonomousSubs: z.preprocess(checkboxValueToBoolean, z.boolean().nullish()),
-		strictDeadline: z.preprocess(checkboxValueToBoolean, z.boolean().nullish()),
-		isInvitational: z.preprocess(checkboxValueToBoolean, z.boolean().nullish()),
-		requireInGameNames: z.preprocess(
-			checkboxValueToBoolean,
-			z.boolean().nullish(),
-		),
-		minMembersPerTeam: z.coerce.number().int().min(1).max(4).nullish(),
-		bracketProgression: bracketProgressionSchema.nullish(),
-	})
-	.refine(
-		async (schema) => {
-			if (schema.eventToEditId) {
-				const eventToEdit = await CalendarRepository.findById({
-					id: schema.eventToEditId,
-				});
-				return schema.date.length === 1 || !eventToEdit?.tournamentId;
-			}
-			return schema.date.length === 1 || !schema.toToolsEnabled;
-		},
-		{
-			message: "Tournament must have exactly one date",
-		},
-	)
-	.refine(
-		(schema) => {
-			if (schema.toToolsMode !== "ALL") {
-				return true;
-			}
-
-			const maps = schema.pool ? MapPool.toDbList(schema.pool) : [];
-
-			return (
-				maps.length === 4 &&
-				rankedModesShort.every((mode) => maps.some((map) => map.mode === mode))
-			);
-		},
-		{
-			message:
-				'Map pool must contain a map for each ranked mode if using "Prepicked by teams - All modes"',
-		},
-	);
